agregados tests de integracion

+++ proyectoIntegrador1.2/src/main/java/com/example/proyectoIntegrador11/repository/OdontologoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OdontologoRepository extends JpaRepository<Odontologo, Integer> {
-    Optional<Odontologo> findByNumeroMatricula(Integer numeroMatricula);
+public interface OdontologoRepository extends JpaRepository<Odontologo, Long> {
+    Optional<Odontologo> findByNumeroMatricula(Long numeroMatricula);
 }
