+++ proyectoIntegrador1.2/src/main/java/com/example/proyectoIntegrador11/repository/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.example.proyectoIntegrador11.repository;
 import com.example.proyectoIntegrador11.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno, Integer> {
+public interface TurnoRepository extends JpaRepository<Turno, Long> {
 }
